查询月租账单接口优化

+++ common/src/main/java/com/jsy/community/entity/CarOrderEntity.java
@@ -69,7 +69,7 @@ public class CarOrderEntity extends BaseEntity {
     /**
      *  月份
      */
-    private Long month;
+    private Integer month;
     /**
      *  1线上支付，2线下支付
      */
